voices en_US, add try_again
copy the voice we have for try_again in en_GB to en_US

It's better than nothing, especially as some people use en_US
as reference to record new voices.

If a native from US can provide a recording with US accent it is still welcome to replace it.
And same if a native from GB can provide one, as the recording we have in en_GB
is not from a native anyway.